requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy
pydantic
requests
python-docx
pdfplumber
werkzeug
psycopg2-binary
gunicorn
pillow 
pytesseract
easyocr
PyJWT
pdf2image